update: legend and title
